add id_s4

+++ student/objdet_pcl.py
@@ -96,7 +96,7 @@ def show_range_image(frame, lidar_name):
 
 
 # create birds-eye view of lidar data
-def bev_from_pcl(lidar_pcl, configs):
+def bev_from_pcl(lidar_pcl, configs, debug=False):
 
     # remove lidar points outside detection area and with too low reflectivity
     mask = np.where((lidar_pcl[:, 0] >= configs.lim_x[0]) & (lidar_pcl[:, 0] <= configs.lim_x[1]) &
@@ -120,7 +120,8 @@ def bev_from_pcl(lidar_pcl, configs):
     # step 3 : perform the same operation as in step 2 for the y-coordinates but make sure that no negative bev-coordinates occur
     lidar_pcl_cpy[:, 1] = np.int_(np.floor(lidar_pcl_cpy[:, 1] / bev_discret) + (configs.bev_width + 1) / 2)
     # step 4 : visualize point-cloud using the function show_pcl from a previous task
-    show_pcl(lidar_pcl_cpy)
+    if debug:
+        show_pcl(lidar_pcl_cpy)
     #######
     ####### ID_S2_EX1 END #######
 
@@ -148,11 +149,12 @@ def bev_from_pcl(lidar_pcl, configs):
     ## step 5 : temporarily visualize the intensity map using OpenCV to make sure that vehicles separate well from the background
     img_intensity = intensity_map * 256
     img_intensity = img_intensity.astype(np.uint8)
-    while (1):
-        cv2.imshow('img_intensity', img_intensity)
-        if cv2.waitKey(10) & 0xFF == 27:
-            break
-    cv2.destroyAllWindows()
+    if debug:
+        while (1):
+            cv2.imshow('img_intensity', img_intensity)
+            if cv2.waitKey(10) & 0xFF == 27:
+                break
+        cv2.destroyAllWindows()
     #######
     ####### ID_S2_EX2 END #######
 
@@ -175,11 +177,12 @@ def bev_from_pcl(lidar_pcl, configs):
     ## step 3 : temporarily visualize the intensity map using OpenCV to make sure that vehicles separate well from the background
     img_height = height_map * 256
     img_height = img_height.astype(np.uint8)
-    while (1):
-        cv2.imshow('img_height', img_height)
-        if cv2.waitKey(10) & 0xFF == 27:
-            break
-    cv2.destroyAllWindows()
+    if debug:
+        while (1):
+            cv2.imshow('img_height', img_height)
+            if cv2.waitKey(10) & 0xFF == 27:
+                break
+        cv2.destroyAllWindows()
     #######
     ####### ID_S2_EX3 END #######
